move docs-site files into same structure as docs
components / styles / tags.
tags export maps of files for reactDocs and reactExample
