+++ packages/docs-site/src/components/colorPalettes.tsx
