+++ packages/docs-site/src/index.tsx
@@ -13,20 +13,15 @@ import * as ReactDOM from "react-dom";
 import {
     createDefaultRenderers,
     Documentation,
-    getTheme,
-    IBaseExampleProps,
     IDocsData,
-    IExampleMap,
     IPackageInfo,
     ReactDocsTagRenderer,
     ReactExampleTagRenderer,
 } from "@blueprintjs/docs";
 
-import * as ReactDocs from "./reactDocs";
+import * as ReactDocs from "./tags/reactDocs";
+import { reactExamples } from "./tags/reactExamples";
 
-import * as CoreExamples from "@blueprintjs/core/examples";
-import * as DateExamples from "@blueprintjs/datetime/examples";
-import * as TableExamples from "@blueprintjs/table/examples";
 
 /* tslint:disable:no-var-requires */
 const docs = require<IDocsData>("./generated/docs.json");
@@ -43,28 +38,8 @@ const versions = require<string[]>("./generated/versions.json")
     } as IPackageInfo));
 /* tslint:enable:no-var-requires */
 
-const SRC_HREF_BASE = "https://github.com/palantir/blueprint/blob/master/packages";
-
 const reactDocs = new ReactDocsTagRenderer(ReactDocs as any);
-
-const examples: IExampleMap = {};
-function addPackageExamples(
-    packageName: string,
-    packageExamples: { [name: string]: React.ComponentClass<IBaseExampleProps> },
-) {
-    for (const exampleName of Object.keys(packageExamples)) {
-        const example = packageExamples[exampleName];
-        const fileName = exampleName.charAt(0).toLowerCase() + exampleName.slice(1) + ".tsx";
-        examples[exampleName] = {
-            render: (props) => React.createElement(example, { ...props, getTheme }),
-            sourceUrl: [SRC_HREF_BASE, packageName, "examples", fileName].join("/"),
-        };
-    }
-}
-addPackageExamples("core", CoreExamples as any);
-addPackageExamples("datetime", DateExamples as any);
-addPackageExamples("table", TableExamples as any);
-const reactExample = new ReactExampleTagRenderer(examples);
+const reactExample = new ReactExampleTagRenderer(reactExamples);
 
 const TAGS = {
     ...createDefaultRenderers(docs),
