+++ packages/docs-site/src/styles/_icons.scss
