+++ packages/docs-site/src/styles/_resources.scss
