+++ packages/docs-site/src/tags/reactExamples.ts
@@ -1,0 +1,28 @@
+import { getTheme, IBaseExampleProps, IExampleMap } from "@blueprintjs/docs";
+import * as React from "react";
+
+import * as CoreExamples from "@blueprintjs/core/examples";
+import * as DateExamples from "@blueprintjs/datetime/examples";
+import * as TableExamples from "@blueprintjs/table/examples";
+
+const SRC_HREF_BASE = "https://github.com/palantir/blueprint/blob/master/packages";
+
+export const reactExamples: IExampleMap = {};
+
+function addPackageExamples(
+    packageName: string,
+    packageExamples: { [name: string]: React.ComponentClass<IBaseExampleProps> },
+) {
+    for (const exampleName of Object.keys(packageExamples)) {
+        const example = packageExamples[exampleName];
+        const fileName = exampleName.charAt(0).toLowerCase() + exampleName.slice(1) + ".tsx";
+        reactExamples[exampleName] = {
+            render: (props) => React.createElement(example, { ...props, getTheme }),
+            sourceUrl: [SRC_HREF_BASE, packageName, "examples", fileName].join("/"),
+        };
+    }
+}
+
+addPackageExamples("core", CoreExamples as any);
+addPackageExamples("datetime", DateExamples as any);
+addPackageExamples("table", TableExamples as any);
